Tests: Move API tests to new infrastructure
